refactor: extract EntryPersister, RelaySlotManager, OrisEntryClient from EntrySportEvent
Business logic extracted to app/Services/SportEvents/Entries/:
- EntryPersister::persist() — UserEntry creation with all fields
- RelaySlotManager::reserveSlot/releaseSlot/availableSlots() — relay slot management with pessimistic locking
- OrisEntryClient::createEntry/deleteEntry() — wraps GuzzleClient ORIS API calls

EntrySportEvent delegates to these services; all 23 characterization
tests updated to call services directly (no more Reflection for these).
